Fix cache issues with document service [fixes luomus/laji#609]
The documents can't be cached by creator, because the personal clause
says "is creator or is in editors", and it didn't take editors into
account.

Editors is an array, so it can't be used as a primary key. Hence, we
disable caching for queries that use the personal clause.

Some bugs were also fixed in the caching system.